AWS Resource Monitor Bug Fix
When I changed the code to return JUnit XML, I changed the logic to no longer use a filter on the state, so that we can track the instances that were running, and shut down. However, the conditional logic was incorrect. I corrected the logic to properly grab the state value.